@@ -3,12 +3,14 @@ package org.egov.servicerequest.web.controllers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.contract.response.ResponseInfo;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.servicerequest.service.ServiceRequestService;
 import org.egov.servicerequest.util.ResponseInfoFactory;
 import org.egov.servicerequest.web.models.Service;
 import org.egov.servicerequest.web.models.ServiceRequest;
 import org.egov.servicerequest.web.models.ServiceResponse;
 import org.egov.servicerequest.web.models.ServiceSearchRequest;
+import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,9 +19,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.Collections;
 import java.util.List;
+
+import static org.egov.servicerequest.error.ErrorCode.INVALID_TENANT_ID_ERR_CODE;
 
 @Slf4j
 @RestController
@@ -37,7 +41,12 @@ public class ServiceController {
 
     @RequestMapping(value="/v1/_create", method = RequestMethod.POST)
     public ResponseEntity<ServiceResponse> create(@RequestBody @Valid ServiceRequest serviceRequest) {
-        Service service = serviceRequestService.createService(serviceRequest);
+        Service service = null;
+        try {
+            service = serviceRequestService.createService(serviceRequest);
+        } catch (InvalidTenantIdException e) {
+            throw new CustomException(INVALID_TENANT_ID_ERR_CODE, e.getMessage());
+        }
         ResponseInfo responseInfo = responseInfoFactory.createResponseInfoFromRequestInfo(serviceRequest.getRequestInfo(), true);
         ServiceResponse response = ServiceResponse.builder().service(Collections.singletonList(service)).responseInfo(responseInfo).build();
         return new ResponseEntity<>(response, HttpStatus.OK);
@@ -45,7 +54,12 @@ public class ServiceController {
 
     @RequestMapping(value="/v1/_search", method = RequestMethod.POST)
     public ResponseEntity<ServiceResponse> search(@Valid @RequestBody ServiceSearchRequest serviceSearchRequest) {
-        List<Service> serviceList = serviceRequestService.searchService(serviceSearchRequest);
+        List<Service> serviceList = null;
+        try {
+            serviceList = serviceRequestService.searchService(serviceSearchRequest);
+        } catch (InvalidTenantIdException e) {
+            throw new CustomException(INVALID_TENANT_ID_ERR_CODE, e.getMessage());
+        }
         ServiceResponse response  = ServiceResponse.builder().service(serviceList).build();
         return new ResponseEntity<>(response,HttpStatus.OK);
     }
