@@ -1,12 +1,14 @@
 package org.egov.servicerequest;
 
 
+import org.egov.common.utils.MultiStateInstanceUtil;
 import org.egov.tracer.config.TracerConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Import;
 
-@Import({ TracerConfiguration.class })
+@Import({ TracerConfiguration.class, MultiStateInstanceUtil.class })
 @SpringBootApplication
 public class ServiceRequestApplication {
 
