@@ -1,6 +1,8 @@
 package org.egov.servicerequest.repository;
 
 import lombok.extern.slf4j.Slf4j;
+import org.egov.common.exception.InvalidTenantIdException;
+import org.egov.common.utils.MultiStateInstanceUtil;
 import org.egov.servicerequest.repository.querybuilder.ServiceDefinitionQueryBuilder;
 import org.egov.servicerequest.repository.rowmapper.ServiceDefinitionRowMapper;
 import org.egov.servicerequest.web.models.Service;
@@ -8,6 +10,7 @@ import org.egov.servicerequest.web.models.ServiceCriteria;
 import org.egov.servicerequest.web.models.ServiceDefinition;
 import org.egov.servicerequest.web.models.ServiceDefinitionCriteria;
 import org.egov.servicerequest.web.models.ServiceDefinitionSearchRequest;
+import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.SingleColumnRowMapper;
@@ -26,13 +29,16 @@ public class ServiceDefinitionRequestRepository {
     private JdbcTemplate jdbcTemplate;
 
     @Autowired
+    private MultiStateInstanceUtil multiStateInstanceUtil;
+
+    @Autowired
     private ServiceDefinitionRowMapper serviceDefinitionRowMapper;
 
     @Autowired
     private ServiceDefinitionQueryBuilder serviceDefinitionQueryBuilder;
 
 
-    public List<ServiceDefinition> getServiceDefinitions(ServiceDefinitionSearchRequest serviceDefinitionSearchRequest) {
+    public List<ServiceDefinition> getServiceDefinitions(ServiceDefinitionSearchRequest serviceDefinitionSearchRequest) throws InvalidTenantIdException {
         ServiceDefinitionCriteria criteria = serviceDefinitionSearchRequest.getServiceDefinitionCriteria();
 
         List<Object> preparedStmtList = new ArrayList<>();
@@ -44,9 +50,11 @@ public class ServiceDefinitionRequestRepository {
         if(CollectionUtils.isEmpty(criteria.getIds())){
             // Fetch ids according to given criteria
             String idQuery = serviceDefinitionQueryBuilder.getServiceDefinitionsIdsQuery(serviceDefinitionSearchRequest, preparedStmtList);
-            log.info("Service definition ids query: " + idQuery);
-            log.info("Parameters: " + preparedStmtList.toString());
+            idQuery = multiStateInstanceUtil.replaceSchemaPlaceholder(idQuery, criteria.getTenantId());
+            log.info("Service definition ids query: {}", idQuery);
+            log.info("Parameters: {}", preparedStmtList);
             List<String> serviceDefinitionIds = jdbcTemplate.query(idQuery, preparedStmtList.toArray(), new SingleColumnRowMapper<>(String.class));
+
 
             if(CollectionUtils.isEmpty(serviceDefinitionIds))
                 return new ArrayList<>();
@@ -59,7 +67,8 @@ public class ServiceDefinitionRequestRepository {
 
 
         String query = serviceDefinitionQueryBuilder.getServiceDefinitionSearchQuery(criteria, preparedStmtList);
-        log.info("query for search: " + query + " params: " + preparedStmtList);
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, criteria.getTenantId());
+        log.info("query for search: {} params: {}", query, preparedStmtList);
         return jdbcTemplate.query(query, preparedStmtList.toArray(), serviceDefinitionRowMapper);
 
     }
