@@ -1,5 +1,6 @@
 package org.egov.servicerequest.validators;
 
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.servicerequest.repository.ServiceDefinitionRequestRepository;
 import org.egov.servicerequest.repository.ServiceRequestRepository;
 import org.egov.servicerequest.web.models.*;
@@ -27,7 +28,7 @@ public class ServiceDefinitionRequestValidator {
     private ServiceRequestRepository serviceRequestRepository;
 
 
-    public void validateServiceDefinitionRequest(ServiceDefinitionRequest serviceDefinitionRequest){
+    public void validateServiceDefinitionRequest(ServiceDefinitionRequest serviceDefinitionRequest) throws InvalidTenantIdException {
         ServiceDefinition serviceDefinition = serviceDefinitionRequest.getServiceDefinition();
 
         // Validate if a service definition with the same combination of tenantId and code already exists
@@ -80,14 +81,14 @@ public class ServiceDefinitionRequestValidator {
         });
     }
 
-    private void validateServiceDefinitionExistence(ServiceDefinition serviceDefinition) {
+    private void validateServiceDefinitionExistence(ServiceDefinition serviceDefinition) throws InvalidTenantIdException {
         List<ServiceDefinition> serviceDefinitionList = serviceDefinitionRequestRepository.getServiceDefinitions(ServiceDefinitionSearchRequest.builder().includeDeleted(true).serviceDefinitionCriteria(ServiceDefinitionCriteria.builder().tenantId(serviceDefinition.getTenantId()).code(Collections.singletonList(serviceDefinition.getCode())).build()).build());
         if(!CollectionUtils.isEmpty(serviceDefinitionList)){
             throw new CustomException(SERVICE_DEFINITION_ALREADY_EXISTS_ERR_CODE, SERVICE_DEFINITION_ALREADY_EXISTS_ERR_MSG);
         }
     }
 
-    private List<ServiceDefinition> validateExistence(ServiceDefinition serviceDefinition) {
+    private List<ServiceDefinition> validateExistence(ServiceDefinition serviceDefinition) throws InvalidTenantIdException {
         List<ServiceDefinition> serviceDefinitionList = serviceDefinitionRequestRepository.
           getServiceDefinitions(ServiceDefinitionSearchRequest.builder()
             .includeDeleted(true)
@@ -101,7 +102,7 @@ public class ServiceDefinitionRequestValidator {
         return serviceDefinitionList;
     }
 
-    private void validateService(List<ServiceDefinition> serviceDefinition){
+    private void validateService(List<ServiceDefinition> serviceDefinition) throws InvalidTenantIdException {
         List<Service> service = serviceRequestRepository.getService(
           ServiceSearchRequest.builder().serviceCriteria(
             ServiceCriteria.builder().serviceDefIds(Collections.singletonList(serviceDefinition.get(0).getId())).build()
@@ -113,7 +114,7 @@ public class ServiceDefinitionRequestValidator {
         }
     }
 
-    public ServiceDefinition validateUpdateRequest(ServiceDefinitionRequest serviceDefinitionRequest) {
+    public ServiceDefinition validateUpdateRequest(ServiceDefinitionRequest serviceDefinitionRequest) throws InvalidTenantIdException {
         ServiceDefinition serviceDefinition = serviceDefinitionRequest.getServiceDefinition();
 
         //Validate if a  Service Definition exists
